requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
httpx>=0.27
pydantic-settings>=2.5
cachetools>=5.5
tenacity>=9.0
aiosqlite>=0.20

pytest>=8.3
pytest-asyncio>=0.24
